Refinery::Core::UsersManager has been moved to core folder
