+++ core/lib/refinery/core/users_manager.rb
